+++ functions.hpp
@@ -1,0 +1,11 @@
+#ifndef FUCNTIONS_HPP
+#define FUNCTIONS_HPP
+
+#include <string>
+
+int countCharacter(std::string s, std::string c);
+void printChoices(std::string choices);
+std::string findNextLabelDirection(std::string choices, int reponse);
+int testReadline(void);
+
+#endif /* FUCNTIONS_HPP */
