+++ main.cpp
@@ -7,7 +7,7 @@
 //
 
 #include <iostream>
-#include "Parsing.cpp"
+#include "functions.hpp"
 
 int main(void) {
     testReadline();
